debug

+++ CS3343_Project_Pilot/src/cs2115Handler/XORHandler.java
@@ -6,7 +6,7 @@ public class XORHandler extends LogicHandler{
 
 	@Override
 	public double handleForDoubleReturn(String expression) throws InvalidExpressionException {
-		return new NOTHandler().handleForDoubleReturn(""+new XNORHandler().handleForDoubleReturn(expression));
+		return new NOTHandler().handleForDoubleReturn(""+(int)new XNORHandler().handleForDoubleReturn(expression));
 	}
 
 }
